Adjustment of p-Values

Adjustment of p-Values

Methods

MultipleTesting.adjust — Function.

adjust(PValues, PValueAdjustment)
adjust(PValues, Int, PValueAdjustment)

Adjustment of p-values

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, BenjaminiHochberg())
4-element Array{Float64,1}:
 0.004
 0.02
 0.04
 0.5
julia> adjust(pvals, 6, BenjaminiHochberg()) # 4 out of 6 p-values
4-element Array{Float64,1}:
 0.006
 0.03
 0.06
 0.75
julia> adjust(pvals, BarberCandes())
4-element Array{Float64,1}:
 0.333333
 0.333333
 0.333333
 1.0

julia> subtypes(PValueAdjustment)
11-element Array{Union{DataType, UnionAll},1}:
 MultipleTesting.BarberCandes
 MultipleTesting.BenjaminiHochberg
 MultipleTesting.BenjaminiHochbergAdaptive
 MultipleTesting.BenjaminiLiu
 MultipleTesting.BenjaminiYekutieli
 MultipleTesting.Bonferroni
 MultipleTesting.ForwardStop
 MultipleTesting.Hochberg
 MultipleTesting.Holm
 MultipleTesting.Hommel
 MultipleTesting.Sidak

See also

PValueAdjustments:

Bonferroni BenjaminiHochberg BenjaminiHochbergAdaptive BenjaminiYekutieli BenjaminiLiu Hochberg Holm Hommel Sidak ForwardStop BarberCandes

Types

MultipleTesting.Bonferroni — Type.

Bonferroni adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, Bonferroni())
4-element Array{Float64,1}:
 0.004
 0.04
 0.12
 1.0

julia> adjust(pvals, 6, Bonferroni())
4-element Array{Float64,1}:
 0.006
 0.06
 0.18
 1.0

References

Bonferroni, C.E. (1936). Teoria statistica delle classi e calcolo delle probabilita (Libreria internazionale Seeber).

MultipleTesting.BenjaminiHochberg — Type.

Benjamini-Hochberg adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, BenjaminiHochberg())
4-element Array{Float64,1}:
 0.004
 0.02
 0.04
 0.5

julia> adjust(pvals, 6, BenjaminiHochberg())
4-element Array{Float64,1}:
 0.006
 0.03
 0.06
 0.75

References

Benjamini, Y., and Hochberg, Y. (1995). Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing. Journal of the Royal Statistical Society. Series B (Methodological) 57, 289–300.

MultipleTesting.BenjaminiHochbergAdaptive — Type.

Adaptive Benjamini-Hochberg adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, BenjaminiHochbergAdaptive(Oracle(0.5))) # known π₀ of 0.5
4-element Array{Float64,1}:
 0.002
 0.01
 0.02
 0.25

julia> adjust(pvals, BenjaminiHochbergAdaptive(StoreyBootstrap())) # π₀ estimator
4-element Array{Float64,1}:
 0.0
 0.0
 0.0
 0.0

julia> adjust(pvals, 6, BenjaminiHochbergAdaptive(StoreyBootstrap()))
4-element Array{Float64,1}:
 0.0
 0.0
 0.0
 0.0

References

Benjamini, Y., and Hochberg, Y. (1995). Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing. Journal of the Royal Statistical Society. Series B (Methodological) 57, 289–300.

MultipleTesting.BenjaminiYekutieli — Type.

Benjamini-Yekutieli adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, BenjaminiYekutieli())
4-element Array{Float64,1}:
 0.00833333
 0.0416667
 0.0833333
 1.0

julia> adjust(pvals, 6, BenjaminiYekutieli())
4-element Array{Float64,1}:
 0.0147
 0.0735
 0.147
 1.0

References

Benjamini, Y., and Yekutieli, D. (2001). The Control of the False Discovery Rate in Multiple Testing under Dependency. The Annals of Statistics 29, 1165–1188.

MultipleTesting.BenjaminiLiu — Type.

Benjamini-Liu adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, BenjaminiLiu())
4-element Array{Float64,1}:
 0.003994
 0.0222757
 0.02955
 0.125

julia> adjust(pvals, 6, BenjaminiLiu())
4-element Array{Float64,1}:
 0.00598502
 0.0408416
 0.0764715
 0.4375

References

Benjamini, Y., and Liu, W. (1999). A step-down multiple hypotheses testing procedure that controls the false discovery rate under independence. Journal of Statistical Planning and Inference 82, 163–170.

MultipleTesting.Hochberg — Type.

Hochberg adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, Hochberg())
4-element Array{Float64,1}:
 0.004
 0.03
 0.06
 0.5

julia> adjust(pvals, 6, Hochberg())
4-element Array{Float64,1}:
 0.006
 0.05
 0.12
 1.0

References

Hochberg, Y. (1988). A sharper Bonferroni procedure for multiple tests of significance. Biometrika 75, 800–802.

MultipleTesting.Holm — Type.

Holm adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, Holm())
4-element Array{Float64,1}:
 0.004
 0.03
 0.06
 0.5

julia> adjust(pvals, 6, Holm())
4-element Array{Float64,1}:
 0.006
 0.05
 0.12
 1.0

References

Holm, S. (1979). A Simple Sequentially Rejective Multiple Test Procedure. Scandinavian Journal of Statistics 6, 65–70.

MultipleTesting.Hommel — Type.

Hommel adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, Hommel())
4-element Array{Float64,1}:
 0.004
 0.03
 0.06
 0.5

julia> adjust(pvals, 6, Hommel())
4-element Array{Float64,1}:
 0.006
 0.05
 0.12
 1.0

References

Hommel, G. (1988). A stagewise rejective multiple test procedure based on a modified Bonferroni test. Biometrika 75, 383–386.

MultipleTesting.Sidak — Type.

Šidák adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, Sidak())
4-element Array{Float64,1}:
 0.003994
 0.039404
 0.114707
 0.9375

julia> adjust(pvals, 6, Sidak())
4-element Array{Float64,1}:
 0.00598502
 0.0585199
 0.167028
 0.984375

References

Šidák, Z. (1967). Rectangular Confidence Regions for the Means of Multivariate Normal Distributions. Journal of the American Statistical Association 62, 626–633.

MultipleTesting.ForwardStop — Type.

Forward-Stop adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, ForwardStop())
4-element Array{Float64,1}:
 0.0010005
 0.00552542
 0.0138367
 0.183664

julia> adjust(pvals, 6, ForwardStop())
4-element Array{Float64,1}:
 0.0010005
 0.00552542
 0.0138367
 0.183664

References

G’Sell, M.G., Wager, S., Chouldechova, A., and Tibshirani, R. (2016). Sequential selection procedures and false discovery rate control. J. R. Stat. Soc. B 78, 423–444.

MultipleTesting.BarberCandes — Type.

Barber-Candès adjustment

Examples

julia> pvals = PValues([0.001, 0.01, 0.03, 0.5]);

julia> adjust(pvals, BarberCandes())
4-element Array{Float64,1}:
 0.333333
 0.333333
 0.333333
 1.0

References

Barber, R.F., and Candès, E.J. (2015). Controlling the false discovery rate via knockoffs. Ann. Statist. 43, 2055–2085.

Arias-Castro, E., and Chen, S. (2017). Distribution-free multiple testing. Electron. J. Statist. 11, 1983–2001.