Combination of p-Values

Combination of p-Values

Methods

MultipleTesting.combine — Function.

combine(PValues, PValueCombination)

Combine p-values

Examples

julia> pvals = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pvals, FisherCombination())
0.007616871850449092
julia> combine(pvals, StoufferCombination())
0.007098326181265917

julia> subtypes(PValueCombination)
7-element Array{Union{DataType, UnionAll},1}:
 MultipleTesting.FisherCombination
 MultipleTesting.LogitCombination
 MultipleTesting.MinimumCombination
 MultipleTesting.SimesCombination
 MultipleTesting.StoufferCombination
 MultipleTesting.TippettCombination
 MultipleTesting.WilkinsonCombination

See also

PValueCombinations:

FisherCombination LogitCombination StoufferCombination TippettCombination SimesCombination WilkinsonCombination MinimumCombination

Types

MultipleTesting.FisherCombination — Type.

Fisher's p-value combination

Examples

julia> pvals = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pvals, FisherCombination())
0.007616871850449092

References

Fisher, R.A. (1925). Statistical methods for research workers (Genesis Publishing Pvt Ltd).

MultipleTesting.LogitCombination — Type.

Logit p-value combination

Examples

julia> pvals = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pvals, LogitCombination())
0.006434494635148462

References

Mudholkar, G.S., and George, E.O. (1977). The Logit Statistic for Combining Probabilities - An Overview (Rochester University NY, Dept of Statistics).

MultipleTesting.StoufferCombination — Type.

Stouffer's p-value combination

Examples

julia> pvals = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pvals, StoufferCombination())
0.007098326181265917

julia> weights = [1.0, 2.0, 0.4, 1.5];

julia> combine(pvals, weights, StoufferCombination())
0.007331653763696742

References

Stouffer, S.A. (1949). The American soldier. Vol. 1: Adjustment during army life (Princeton University Press).

Liptak, T. (1958). On the combination of independent tests. Magyar Tud Akad Mat Kutato Int Kozl 3, 171–197.

MultipleTesting.TippettCombination — Type.

Tippett's p-value combination

Examples

julia> pvals = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pvals, TippettCombination())
0.039403990000000055

References

Tippett, L.H.C. (1931). The Methods of Statistics. An introduction mainly for workers in the biological sciences.

MultipleTesting.SimesCombination — Type.

Simes's p-value combination

Examples

julia> pvals = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pvals, SimesCombination())
0.04

References

Simes, R.J. (1986). An improved Bonferroni procedure for multiple tests of significance. Biometrika 73, 751–754.

MultipleTesting.WilkinsonCombination — Type.

Wilkinson's p-value combination

Examples

julia> pv = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pv, WilkinsonCombination(1))  # combination with rank 1
0.03940399000000003

julia> combine(pv, WilkinsonCombination(4))  # combination with rank 4
0.0625

References

Wilkinson, B. (1951). A statistical consideration in psychological research. Psychological Bulletin 48, 156.

MultipleTesting.MinimumCombination — Type.

Minimum of adjusted p-value combination

Examples

julia> pv = PValues([0.01, 0.02, 0.3, 0.5]);

julia> combine(pv, MinimumCombination(BenjaminiHochberg()))
0.04

julia> combine(pv, MinimumCombination(ForwardStop()))
0.01005033585350145